+++ packages/wrapper-react/test/index.extapi.test.tsx
@@ -100,13 +100,13 @@ describe.sequential('Test MonacoEditorReactComp: External monaco-vscode-api', ()
         renderResult.rerender(<MonacoEditorReactComp
             editorAppConfig={editorAppConfig2}
             style={{ 'height': '800px' }}
-            toggleReprocessConfig={true}
-            onConfigProcessed={async (result, editorApp?: EditorApp) => {
+            triggerReprocessConfig={1}
+            onConfigProcessed={async (result) => {
                 expect(result.textUpdated).toBeTruthy();
                 expect(result.modelUpdated).toBeFalsy();
-                expect(editorApp).toBeDefined();
-                expect(editorApp?.getEditor()?.getValue()).toBe(codeUpdated);
-                expect(editorApp?.getTextModels().modified?.getValue()).toBe(codeUpdated);
+                expect(result.editorApp).toBeDefined();
+                expect(result.editorApp?.getEditor()?.getValue()).toBe(codeUpdated);
+                expect(result.editorApp?.getTextModels().modified?.getValue()).toBe(codeUpdated);
                 deferred2.resolve();
             }}
         />);
@@ -202,7 +202,7 @@ describe.sequential('Test MonacoEditorReactComp: External monaco-vscode-api', ()
 
         const App = () => {
             const [codeState, setCodeState] = useState<string>(code);
-            const [reprocessConfig, setReprocessConfig] = useState<boolean>(false);
+            const [triggerReprocessConfig, setTriggerReprocessConfig] = useState<number>(0);
 
             const editorAppConfig = createDefaultEditorAppConfig({
                 modified: {
@@ -215,7 +215,7 @@ describe.sequential('Test MonacoEditorReactComp: External monaco-vscode-api', ()
                 <>
                     <button id='change-button' style={{background: 'purple'}} onClick={() => {
                         setCodeState(codeUpdated);
-                        setReprocessConfig(!reprocessConfig);
+                        setTriggerReprocessConfig(triggerReprocessConfig + 1);
                     }}>Change Text</button>
                     <MonacoEditorReactComp
                         editorAppConfig={editorAppConfig}
@@ -228,7 +228,7 @@ describe.sequential('Test MonacoEditorReactComp: External monaco-vscode-api', ()
                                 deferredChanged.resolve();
                             }
                         }}
-                        toggleReprocessConfig={reprocessConfig}
+                        triggerReprocessConfig={triggerReprocessConfig}
                         onConfigProcessed={(result) => {
                             expect(result.textUpdated).toBeTruthy();
                             expect(result.modelUpdated).toBeFalsy();