+++ packages/wrapper-react/test/index.lc.test.tsx
@@ -6,7 +6,6 @@
 import { render } from '@testing-library/react';
 import { MonacoEditorReactComp } from '@typefox/monaco-editor-react';
 import { Deferred, delayExecution } from 'monaco-languageclient/common';
-import type { EditorApp } from 'monaco-languageclient/editorApp';
 import type { LanguageClientManager } from 'monaco-languageclient/lcwrapper';
 import type { MonacoVscodeApiConfig } from 'monaco-languageclient/vscodeApiWrapper';
 import React from 'react';
@@ -135,12 +134,12 @@ describe.sequential('Test MonacoEditorReactComp: Langugae Client', () => {
             editorAppConfig={editorAppConfig2}
             languageClientConfig={languageClientConfig}
             style={{ 'height': '800px' }}
-            toggleReprocessConfig={true}
-            onConfigProcessed={async (result, editorApp?: EditorApp) => {
+            triggerReprocessConfig={1}
+            onConfigProcessed={async (result) => {
                 expect(result.textUpdated).toBe(true);
-                expect(editorApp).toBeDefined();
+                expect(result.editorApp).toBeDefined();
                 await delayExecution(hundredMs);
-                expect(editorApp?.getEditor()?.getValue()).toBe(codeUpdated);
+                expect(result.editorApp?.getEditor()?.getValue()).toBe(codeUpdated);
                 deferred2.resolve();
             }}
         />);
@@ -194,12 +193,12 @@ describe.sequential('Test MonacoEditorReactComp: Langugae Client', () => {
             editorAppConfig={editorAppConfig2}
             languageClientConfig={languageClientConfigs2}
             style={{ 'height': '800px' }}
-            toggleReprocessConfig={true}
-            onConfigProcessed={async (result, editorApp?: EditorApp) => {
+            triggerReprocessConfig={1}
+            onConfigProcessed={async (result) => {
                 expect(result.textUpdated).toBe(true);
-                expect(editorApp).toBeDefined();
+                expect(result.editorApp).toBeDefined();
                 await delayExecution(hundredMs);
-                expect(editorApp?.getEditor()?.getValue()).toBe(codeUpdated);
+                expect(result.editorApp?.getEditor()?.getValue()).toBe(codeUpdated);
                 deferred2.resolve();
             }}
             onError={(error) => {