+++ packages/wrapper-react/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to npm module [@typefox/monaco-editor-react](https://www.npm
 ## [7.4.0-next.1] - 2025-11-20
 
 - Use flags to trigger behaviour [#997](https://github.com/TypeFox/monaco-languageclient/pull/997).
-- Updated all `@codingame/monaco-vscode` packages to `23.0.4`.
+- Updated all `@codingame/monaco-vscode` packages to `23.1.0`.
 
 ## [7.3.0] - 2025-11-16